Fix prototype args for tparm(3).

+++ src/terminfo.c
@@ -35,7 +35,7 @@ char *
 tparam (string, outstring, len, arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9)
      char *string;
      char *outstring;
-     int arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9;
+     long arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9;
 {
   char *temp;
